+++ src/state/globalState.js
@@ -15,6 +15,7 @@ class GimletGeneralState {
         this.lldbLibrary = this.getLldbLibraryPath();
         this.tcpPort = DEFAULT_TCP_PORT;
         this.stopOnEntry = true;
+        this.sbfTraceDir = null;
     }
 
     // TODO: Implement a mechanism to allow user to pass custom path trough gimlet.json config file.
@@ -62,6 +63,9 @@ class GimletGeneralState {
         }
         if (config.stopOnEntry !== undefined) {
             this.stopOnEntry = config.stopOnEntry;
+        }
+        if (config.sbfTraceDir !== undefined) {
+            this.sbfTraceDir = config.sbfTraceDir;
         }
         if (
             config.platformToolsVersion !== undefined &&